% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plast.R
\name{plast}
\alias{plast}
\title{Plot the last plot}
\usage{
plast(plot.name = "", loc = "/Users/csng/work/rplots/", w = 5, h = 5)
}
\arguments{
\item{plot.name}{plot object}

\item{loc}{output location}

\item{w}{width}

\item{h}{height}
}
\value{
System output pdf
}
\description{
Plot the last plot
}
\examples{
plast()
}
