% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savetab.R
\name{savetab}
\alias{savetab}
\title{Save dataframe}
\usage{
savetab(object, filepath, archive.overwrite = T, ...)
}
\arguments{
\item{object}{dataframe}

\item{filepath}{output}

\item{archive.overwrite}{if TRUE will overwrite and archive to an archive folder}

\item{...}{additional arguments to pass to write.table if you so wish}
}
\value{
saved df
}
\description{
Save dataframe
}
\examples{
savetab(datasets::mtcars,"./mtcars.csv")
}
