% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewExcel.R
\name{ViewExcel}
\alias{ViewExcel}
\title{View dataframe in a csv}
\usage{
ViewExcel(df = .Last.value, file = tempfile(fileext = ".csv"))
}
\arguments{
\item{df}{dataframe}

\item{file}{temporary file name}
}
\value{
temporary csv
}
\description{
View dataframe in a csv
}
\examples{
ViewExcel(datasets::mtcars)
}
